+++ test-crates/uniffi-pure/src/lib.rs
@@ -1,0 +1,5 @@
+fn add(a: u32, b: u32) -> u32 {
+    a + b
+}
+
+uniffi_macros::include_scaffolding!("math");